requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.0
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
playwright>=1.41.0
aiohttp>=3.9.0
python-docx>=1.1.0
ollama>=0.3.0
openai>=1.12.0
sse-starlette>=2.0.0
huggingface-hub>=0.27.0
psutil>=5.9.0
pytest>=8.0.0
# Transformers + CUDA-enabled torch (cu124 bundles cublas64_12.dll, works with CUDA 13.x driver)
transformers>=4.51.0
accelerate>=0.30.0
# Install torch CUDA separately — run this ONCE after setup.ps1:
#   .\\venv\\Scripts\\pip.exe install torch torchvision --index-url https://download.pytorch.org/whl/cu124
torch>=2.4.0
# Optional: tiktoken for accurate token counts when using DeepSeek (pip install tiktoken)
# aiofiles — optional if async file I/O is added

# llama-cpp-python cu124 prebuilt wheel (Python 3.12 Windows):
#   .\\venv\\Scripts\\pip.exe install "https://github.com/abetlen/llama-cpp-python/releases/download/v0.3.4-cu124/llama_cpp_python-0.3.4-cp312-cp312-win_amd64.whl"
# Note: the cu124 wheel requires torch cu124 to be installed first (provides cublas64_12.dll)
llama-cpp-python==0.3.4
